+++ Spotify-API-Practice/AppDelegate.swift
@@ -10,22 +10,24 @@ import UIKit
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
-    var window: UIWindow?
-    private var loginViewController: LoginViewController?
+    //var window: UIWindow?
+    //private var loginViewController: LoginViewController?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         
-        let window = UIWindow(frame: UIScreen.main.bounds)
-        self.window = window
+        //let window = UIWindow(frame: UIScreen.main.bounds)
+        //self.window = window
+        //
+        //let storyboard = UIStoryboard(name: "Login", bundle: nil)
+        //let vc = storyboard.instantiateInitialViewController() as! LoginViewController
+        //let nav = UINavigationController(rootViewController: vc)
+        //
+        //self.loginViewController = vc
+        //
+        //window.rootViewController = nav
+        //window.makeKeyAndVisible()
         
-        let storyboard = UIStoryboard(name: "Login", bundle: nil)
-        let vc = storyboard.instantiateInitialViewController() as! LoginViewController
-        let nav = UINavigationController(rootViewController: vc)
-        
-        self.loginViewController = vc
-        
-        window.rootViewController = nav
-        window.makeKeyAndVisible()
+        Router.shared.showRoot(window: UIWindow(frame: UIScreen.main.bounds))
         
         return true
     }
@@ -33,14 +35,14 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ app: UIApplication, open url: URL, options: [UIApplication.OpenURLOptionsKey : Any] = [:]) -> Bool {
         
         
-        guard let loginViewController = self.loginViewController else {
-          return true
-        }
-        
-        print("AppDelegateに戻ってきた")
-        print("リダイレクトパラメータ(URL)：\(url)")
-        loginViewController.openURL(url: url)
-        
+        //guard let loginViewController = self.loginViewController else {
+        //  return true
+        //}
+        //
+        //print("AppDelegateに戻ってきた")
+        //print("リダイレクトパラメータ(URL)：\(url)")
+        //loginViewController.openURL(url: url)
+        Router.shared.afterRedirect(url: url)
         
         return true
     }
